+++ __tests__/globals.js
@@ -1,0 +1,2 @@
+__BACKEND__     = 'api.project.vm';
+__ENVIRONMENT__ = 'development';