+++ http/gateway.js
@@ -6,7 +6,6 @@ var http         = require('http');
 var https        = require('https');
 var HttpError    = require('./error');
 var Extendable   = require('../lib/extendable');
-var dispatcher   = require('../lib/dispatcher');
 
 var HttpGateway = Extendable.extend({
 
@@ -21,21 +20,24 @@ var HttpGateway = Extendable.extend({
     /**
      * Perform a request to the API
      *
-     * @param  string  method Request method
-     * @param  string  path   Request path
-     * @param  object  data   Request data (if any)
+     * @param  string  method   Request method
+     * @param  string  path     Request path
+     * @param  object  data     Request data (if any)
+     * @param  object  headers  Additional headers (if any)
      * @return promise
      */
-    apiRequest : function(method, path, data)
+    apiRequest : function(method, path, data, headers)
     {
         var gateway = this;
 
-        if (data && method === 'GET') {
-            path = path + '?' + this._toQuery(data);
+        if (_.isUndefined(headers)) {
+            headers = {};
         }
 
-        return Q.Promise(_.bind(function(resolve, reject) {
-            var options  = this._getRequestOptions(method, path);
+        return Q.Promise(function(resolve, reject) {
+            var options = gateway.getRequestOptions(method, path, data);
+
+            _.extend(options.headers, headers);
 
             var req = (gateway.getConfig().secure ? https : http).request(options, function(response) {
                 var responseText = '';
@@ -45,7 +47,7 @@ var HttpGateway = Extendable.extend({
                 });
 
                 response.on('end', function() {
-                    var responseData;
+                    var responseData, shouldReject;
 
                     try {
                         responseData = JSON.parse(responseText);
@@ -54,10 +56,7 @@ var HttpGateway = Extendable.extend({
                     }
 
                     if (response.statusCode >= 400) {
-                        if (response.statusCode === 401) {
-                            dispatcher.emit('401-response-received');
-                        }
-                        reject(new HttpError(responseData, response));
+                        gateway.handleError(response, responseData, resolve, reject, method, path, data, headers);
                     } else {
                         resolve(responseData);
                     }
@@ -77,14 +76,22 @@ var HttpGateway = Extendable.extend({
             }
 
             req.end();
-        }, this));
+        });
     },
 
-    _getRequestOptions : function(method, path)
+    getRequestOptions : function(method, path, data)
     {
-        var config;
+        var config, queryString;
 
         config = this.getConfig();
+
+        if (data && method === 'GET') {
+            queryString = this.toQuery(data);
+
+            if (queryString) {
+                path = path + '?' + queryString;
+            }
+        }
 
         return {
             hostname        : config.hostname,
@@ -105,9 +112,9 @@ var HttpGateway = Extendable.extend({
      * @param  object data Query data in object form
      * @return string
      */
-    _toQuery : function(data)
+    toQuery : function(data)
     {
-        return this._buildParams('', data);
+        return this.buildParams('', data);
     },
 
     /**
@@ -121,7 +128,7 @@ var HttpGateway = Extendable.extend({
      * @param  bool   top    Is this a recursive call to this function?
      * @return string
      */
-    _buildParams : function(prefix, val, top)
+    buildParams : function(prefix, val, top)
     {
         var instance = this;
 
@@ -129,14 +136,14 @@ var HttpGateway = Extendable.extend({
 
         if (_.isArray(val)) {
             return _.map(val, function(value, key) {
-                return instance._buildParams(top ? key : prefix + '[]', value, false);
+                return instance.buildParams(top ? key : prefix + '[]', value, false);
             }).join('&');
         } else if (_.isObject(val)) {
             return _.map(val, function(value, key) {
-                return instance._buildParams(top ? key : prefix + '[' + key + ']', value, false);
+                return instance.buildParams(top ? key : prefix + '[' + key + ']', value, false);
             }).join('&');
         } else {
-            return this._urlEncode(prefix) + '=' + this._urlEncode(val);
+            return this.urlEncode(prefix) + '=' + this.urlEncode(val);
         }
     },
 
@@ -149,8 +156,26 @@ var HttpGateway = Extendable.extend({
      * @param  string string String to encode
      * @return string
      */
-    _urlEncode : function(string) {
+    urlEncode : function(string)
+    {
         return encodeURIComponent(string).replace(/\%20/g, '+');
+    },
+
+    /**
+     * Handle API request errors
+     *
+     * @param  {Object}   response     The API response
+     * @param  {Mixed}    responseData The data returned in the response
+     * @param  {Function} resolve      The success callback
+     * @param  {Function} reject       The fail callback
+     * @param  {String}   method       The failed request's method
+     * @param  {String}   path         The failed request's path
+     * @param  {Object}   data         The failed request body data (if any)
+     * @param  {Object}   headers      The extra headers set on the failed request
+     */
+    handleError : function(response, responseData, resolve, reject, method, path, data, headers)
+    {
+        reject(new HttpError(responseData, response));
     }
 
 });