+++ store/auth.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var BaseStore       = require('./base');
-var HttpAuthGateway = require('../http/auth-gateway');
-
-var AuthStore = BaseStore.extend(HttpAuthGateway.prototype);
-
-module.exports = AuthStore;